+++ src/collective/volto/formsupport/tests/test_send_action_form.py
@@ -11,7 +11,6 @@ from plone.app.testing import TEST_USER_ID
 from plone.registry.interfaces import IRegistry
 from plone.restapi.testing import RelativeSession
 from Products.MailHost.interfaces import IMailHost
-from six import StringIO
 from zope.component import getUtility
 
 import base64
@@ -625,11 +624,11 @@ class TestMailSend(unittest.TestCase):
             # Python 3 with Products.MailHost 4.10+
             msg = msg.decode("utf-8")
 
-        parsed_msg = Parser().parse(StringIO(msg))
+        parsed_msg = Parser().parsestr(msg)
         self.assertEqual(parsed_msg.get("from"), "john@doe.com")
         self.assertEqual(parsed_msg.get("to"), "smith@doe.com")
         self.assertEqual(parsed_msg.get("subject"), "block subject")
-        msg_body = parsed_msg.get_payload(decode=True).decode()
+        msg_body = parsed_msg.get_payload()[1].get_payload().replace("=\r\n", "")
         self.assertIn(
             "<p>This message will be sent to the person filling in the form.</p>",
             msg_body,
@@ -676,11 +675,12 @@ class TestMailSend(unittest.TestCase):
         if isinstance(msg, bytes) and bytes is not str:
             # Python 3 with Products.MailHost 4.10+
             msg = msg.decode("utf-8")
-        parsed_msg = Parser().parse(StringIO(msg))
+        parsed_msg = Parser().parsestr(msg)
         self.assertEqual(parsed_msg.get("from"), "john@doe.com")
         self.assertEqual(parsed_msg.get("to"), "site_addr@plone.com")
         self.assertEqual(parsed_msg.get("subject"), "block subject")
-        msg_body = parsed_msg.get_payload(decode=True).decode()
+
+        msg_body = parsed_msg.get_payload()[1].get_payload().replace("\r\n", "")
         self.assertIn("<strong>Message:</strong> just want to say hi", msg_body)
         self.assertIn("<strong>Name:</strong> Smith", msg_body)
 
@@ -689,11 +689,14 @@ class TestMailSend(unittest.TestCase):
             # Python 3 with Products.MailHost 4.10+
             acknowledgement_message = acknowledgement_message.decode("utf-8")
 
-        parsed_ack_msg = Parser().parse(StringIO(acknowledgement_message))
+        parsed_ack_msg = Parser().parsestr(acknowledgement_message)
         self.assertEqual(parsed_ack_msg.get("from"), "john@doe.com")
         self.assertEqual(parsed_ack_msg.get("to"), "smith@doe.com")
         self.assertEqual(parsed_ack_msg.get("subject"), "block subject")
-        ack_msg_body = parsed_ack_msg.get_payload(decode=True).decode()
+
+        ack_msg_body = (
+            parsed_ack_msg.get_payload()[1].get_payload().replace("=\r\n", "")
+        )
         self.assertIn(
             "<p>This message will be sent to the person filling in the form.</p>",
             ack_msg_body,
@@ -729,6 +732,7 @@ class TestMailSend(unittest.TestCase):
         if isinstance(msg, bytes) and bytes is not str:
             # Python 3 with Products.MailHost 4.10+
             msg = msg.decode("utf-8")
+        msg.replace("\r\n", "")
 
         self.assertIn(f"Subject: {subject}", msg)
         self.assertIn("From: john@doe.com", msg)
@@ -741,24 +745,22 @@ class TestMailSend(unittest.TestCase):
             f"<caption>Form submission data for {self.document.title}</caption>", msg
         )
         self.assertIn(
-            """<thead>
-      <tr role="row">
-        <th align="left" scope="col" role="columnheader">Field</th>
-        <th align="left" scope="col" role="columnheader">Value</th>
-      </tr>
-    </thead>""",
+            """<th align="left" scope="col" role="columnheader">Field</th>""",
+            msg,
+        )
+        self.assertIn(
+            """<th align="left" scope="col" role="columnheader">Value</th>""",
             msg,
         )
 
         self.assertIn(
-            """<tr role="row">
-          <th align="left" scope="row" role="rowheader">Name</th>""",
+            """<th align="left" scope="row" role="rowheader">Name</th>""",
             msg,
         )
+
         self.assertIn(f'<td align="left">{name}</td>', msg)
         self.assertIn(
-            """<tr role="row">
-          <th align="left" scope="row" role="rowheader">""",
+            """<th align="left" scope="row" role="rowheader">""",
             msg,
         )
         self.assertIn(f'<td align="left">{message}</td>', msg)
@@ -821,9 +823,10 @@ class TestMailSend(unittest.TestCase):
             # Python 3 with Products.MailHost 4.10+
             msg = msg.decode("utf-8")
 
-        parsed_msgs = Parser().parse(StringIO(msg))
+        parsed_msgs = Parser().parsestr(msg)
         # 1st index is the XML attachment
         msg_contents = parsed_msgs.get_payload()[1].get_payload(decode=True)
+
         xml_tree = ET.fromstring(msg_contents)
         for index, field in enumerate(xml_tree):
             self.assertEqual(field.get("name"), form_data[index]["label"])